+++ dacite/src/core/command_buffer.rs
@@ -39,6 +39,11 @@ impl VulkanObject for CommandBuffer {
     type NativeVulkanObject = vks::VkCommandBuffer;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle()
     }