+++ dacite/src/core/descriptor_set_layout.rs
@@ -29,6 +29,11 @@ impl VulkanObject for DescriptorSetLayout {
     type NativeVulkanObject = vks::VkDescriptorSetLayout;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle()
     }