+++ dacite/src/core/device.rs
@@ -53,6 +53,11 @@ impl VulkanObject for Device {
     type NativeVulkanObject = vks::VkDevice;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle()
     }