+++ dacite/src/core/instance.rs
@@ -87,6 +87,11 @@ impl VulkanObject for Instance {
     type NativeVulkanObject = vks::VkInstance;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle()
     }