+++ dacite/src/core/physical_device.rs
@@ -76,6 +76,11 @@ impl VulkanObject for PhysicalDevice {
     type NativeVulkanObject = vks::VkPhysicalDevice;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle
     }