+++ dacite/src/core/queue.rs
@@ -65,6 +65,11 @@ impl VulkanObject for Queue {
     type NativeVulkanObject = vks::VkQueue;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle
     }