+++ dacite/src/ext_debug_report/debug_report_callback.rs
@@ -30,6 +30,11 @@ impl VulkanObject for DebugReportCallbackExt {
     type NativeVulkanObject = vks::VkDebugReportCallbackEXT;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle()
     }