+++ dacite/src/khr_display/display.rs
@@ -66,6 +66,11 @@ impl VulkanObject for DisplayKhr {
     type NativeVulkanObject = vks::VkDisplayKHR;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle
     }