+++ dacite/src/khr_display/display_mode.rs
@@ -65,6 +65,11 @@ impl VulkanObject for DisplayModeKhr {
     type NativeVulkanObject = vks::VkDisplayModeKHR;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle
     }