+++ dacite/src/khr_swapchain/swapchain.rs
@@ -30,6 +30,11 @@ impl VulkanObject for SwapchainKhr {
     type NativeVulkanObject = vks::VkSwapchainKHR;
 
     #[inline]
+    fn id(&self) -> u64 {
+        self.as_native_vulkan_object() as u64
+    }
+
+    #[inline]
     fn as_native_vulkan_object(&self) -> Self::NativeVulkanObject {
         self.handle()
     }